(#532)  code  update  register  interface

+++ app/src/androidTest/java/com/zhiyicx/thinksnsplus/comment/CommentTest.java
@@ -95,7 +95,7 @@ public class CommentTest {
         String comment_mark = AppApplication.getmCurrentLoginAuth().getUser_id()
                 + "" + System.currentTimeMillis();
         createComment.setComment_mark(Long.parseLong(comment_mark));
-        createComment.setUser_id(AppApplication.getmCurrentLoginAuth().getUser_id());
+        createComment.setUser_id((int) AppApplication.getmCurrentLoginAuth().getUser_id());
         createComment.setCreated_at(TimeUtils.getCurrenZeroTimeStr());
 
         UserInfoBean userInfoBean = new UserInfoBean();
